Extract CapturePreviewStrip.swift
